Re-organizing to ditch ECDSA, OpenSSL and Boost

+++ libzerocoin/ZVector.h
@@ -12,10 +12,10 @@ std::vector<IntegerMod<T_> > ZVector(
 {
     std::vector<IntegerMod<T_> > Z(m1 + m2 + 1);
     for (size_t i = 0; i < m1; i++)
-        Z[i] = (x ^ ((i - m1) * n));
+        Z[i] = (x ^ int64_t((i - m1) * n));
     size_t k = 0;
     for (size_t i = m1; i < m1 + m2; i++) {
-        Z[i] = (x ^ (k * n + 1));
+        Z[i] = (x ^ int64_t(k * n + 1));
         k++;
     }
     Z[m1 + m2] = x ^ 2;
